Update plotting.py
In line 161, change the centre coordinates datatype of circle function in cv2 module into int type.

+++ img/plotting.py
@@ -158,7 +158,7 @@ def annotate(image, target, classes, display_boxes=True, display_classes=True, c
             filtered = False
             for p in points:
                 if p[2] != constants.KP_FILTERED:
-                    cv2.circle(image, (p[0], p[1]), int(image_dim/100), color=color, thickness=-thickness)
+                    cv2.circle(image, (int(p[0]), int(p[1])), int(image_dim/100), color=color, thickness=-thickness)
                 else:
                     cv2.circle(image, (p[0], p[1]), int(image_dim/100), color=(128/255, 128/255, 128/255),
                                thickness=-thickness)
